fix(graph): pinned_slot 엔티티 매핑 Integer→Short — 기동 스키마 검증 오류 해소
V44가 pinned_slot을 SMALLINT(int2)로 생성했으나 엔티티가 Integer(int4)로
매핑돼 Hibernate ddl-auto=validate가 기동 시 거부. 값이 1~5라 SMALLINT가
적절하므로 마이그레이션은 유지하고 필드를 Short로 변경해 매핑 일치.
gradlew test는 실 postgres validate를 안 거쳐 미검출(B-12).

+++ backend/src/main/java/com/codeprint/domain/graph/Graph.java
@@ -37,8 +37,9 @@ public class Graph {
     @Column(name = "total_file_count")
     private Integer totalFileCount;
 
+    // SMALLINT(int2) 컬럼과 매핑 일치를 위해 Short — 슬롯 값은 1~5
     @Column(name = "pinned_slot")
-    private Integer pinnedSlot;
+    private Short pinnedSlot;
 
     // 프로젝트 ID와 분석 ID로 새 그래프 인스턴스 생성
     public static Graph create(UUID projectId, UUID analysisId) {
@@ -67,7 +68,7 @@ public class Graph {
         if (slot < 1 || slot > 5) {
             throw new IllegalArgumentException("고정 슬롯은 1~5만 허용됩니다: " + slot);
         }
-        this.pinnedSlot = slot;
+        this.pinnedSlot = (short) slot;
         this.updatedAt = Instant.now();
     }
